Fixed bugs in benchmarks

+++ src/main/kotlin/com/lsmt/core/Types.kt
@@ -7,7 +7,7 @@ typealias Compactor = Runnable
 typealias Record = ByteArray
 typealias Entry = Pair<String, Record?>
 typealias LevelIndex = SortedMap<Int, Level>
-typealias NumberedFile = Pair<Int, Path>
+typealias NumberedFile = Pair<Long, Path>
 
 class KeyRange(override val start: String, override val endInclusive: String) : ClosedRange<String>
